iterative

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -25,7 +25,7 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 	{
 		return;
 	}
-	for (size_t i = 0; i < size; i++)
+	for (unsigned int i = 0; i < size; i++)
 	{
 		(*action)(array[i]);
 	}
